+++ pre-eval/evaluation_eqdis/config.py
@@ -49,6 +49,14 @@ def read_config():
                 for i in item[1].split(','):
                     tmp.append(i)
                 config_info[key] = tmp
+            elif key == 'lonBegin':
+                config_info[key] = item[1]
+            elif key == 'lonEnd':
+                config_info[key] = item[1]
+            elif key == 'latBegin':
+                config_info[key] = item[1]
+            elif key == 'latEnd':
+                config_info[key] = item[1]
             else:
                 print('no this item: {}'.format(key))
                 assert False
