+++ pre-eval/pytorch_pre1.4/NpyConvertNc.py
@@ -30,8 +30,7 @@ def createDistanceNc(config_dict, ncPath ='test.nc'):
     # 获取开始时间
     start_time = datetime.datetime.strptime(config_dict['Datetime'],'%Y%m%d%H%M')
     # 获取结束时间
-    end_time = datetime.datetime.strptime(config_dict['Datetime'],'%Y%m%d%H%M') + datetime.timedelta(hours=config_dict['ForecastHourNum'])
-
+    end_time = datetime.datetime.strptime(config_dict['Datetime'], '%Y%m%d%H%M') + datetime.timedelta(hours=config_dict['ForecastHourNum'])
 
 
     f_w.variables['Flash_pre'].init_time = start_time.strftime('%Y%m%d') + '_' + start_time.strftime('%H%M%S')
